rename QrResolver Hook
