+++ react/src/components/QrReader/QrReaderContainer.tsx
@@ -1,6 +1,6 @@
 import { useBoolean } from "@chakra-ui/react";
 import { useErrorHandling } from "hooks/useErrorHandling";
-import { IQrResolvedValue, IQrResolverResultKind, useQrResolver } from "hooks/QrResolver";
+import { IQrResolvedValue, IQrResolverResultKind, useQrResolver } from "hooks/useQrResolver";
 import { useCallback } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import QrReader from "./components/QrReader";
