+++ react/src/hooks/useQrResolver.ts
